pyproject.toml
[project]
name = "tts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-whisper>=20240930",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.11.12",
]
